requirements.txt
numpy
pandas
matplotlib
openai
fpdf
python-pptx
